==> analysis/01_ImportData_LoadPackages.R <==

# Load Packages Required --------------------------------------------------



library(tidyverse)


# Import Dataset ----------------------------------------------------------

mktgCmpgnUnProcessed <- read_csv("input/marketing_campaign.csv")
